src/sciforge/physics/fluids.py
"""
Fluid dynamics module implementing various fluid simulations
"""

import numpy as np
from typing import Optional, Tuple, List
from .base import PhysicalSystem
from ..numerical.integration import trapezoid
from ..core.constants import CONSTANTS

class FluidColumn(PhysicalSystem):
    """Class representing a fluid column for Plateau-Rayleigh instability simulation"""
    
    def __init__(self,
                 radius: float,
                 length: float,
                 density: float,
                 surface_tension: float,
                 viscosity: float,
                 n_points: int = 100):
        """
        Initialize fluid column
        
        Args:
            radius: Initial radius of fluid column (m)
            length: Length of fluid column (m)
            density: Fluid density (kg/m³)
            surface_tension: Surface tension coefficient (N/m)
            viscosity: Dynamic viscosity (Pa·s)
            n_points: Number of discretization points
        """
        super().__init__(mass=density * np.pi * radius**2 * length,
                        position=np.zeros(3))
        
        self.radius = radius
        self.length = length
        self.density = density
        self.surface_tension = surface_tension
        self.viscosity = viscosity
        
        # Discretize the column
        self.z = np.linspace(0, length, n_points)
        self.r = np.ones_like(self.z) * radius
        self.v_r = np.zeros_like(self.z)  # Radial velocity
        
        # Store history
        self.history = {
            'time': [0.0],
            'radius': [self.r.copy()],
            'velocity': [self.v_r.copy()]
        }
        
    def calculate_pressure(self) -> np.ndarray:
        """Calculate pressure due to surface tension"""
        # Curvature terms
        d2r_dz2 = np.gradient(np.gradient(self.r, self.z), self.z)
        axial_term = d2r_dz2 / (1 + np.gradient(self.r, self.z)**2)**(3/2)
        radial_term = 1 / self.r
        
        # Laplace pressure
        return self.surface_tension * (radial_term - axial_term)
    
    def calculate_growth_rate(self) -> float:
        """Calculate theoretical growth rate of fastest growing mode"""
        k = 0.697 / self.radius  # Wavenumber of fastest growing mode
        omega = np.sqrt(self.surface_tension / (self.density * self.radius**3))
        return omega
    
    def update(self, dt: float):
        """Update fluid column state"""
        # Calculate pressure
        pressure = self.calculate_pressure()
        
        # Update radial velocity using Navier-Stokes
        d2v_dz2 = np.gradient(np.gradient(self.v_r, self.z), self.z)
        acceleration = (-np.gradient(pressure, self.z) / self.density +
                       self.viscosity * d2v_dz2 / self.density)
        
        self.v_r += acceleration * dt
        
        # Update radius
        self.r += self.v_r * dt
        
        # Enforce volume conservation (approximately)
        volume = trapezoid(lambda z: np.pi * self.r**2, 0, self.length, len(self.z))
        scale = np.sqrt(self.length * np.pi * self.radius**2 / volume)
        self.r *= scale
        
        # Update history
        self.history['time'].append(self.history['time'][-1] + dt)
        self.history['radius'].append(self.r.copy())
        self.history['velocity'].append(self.v_r.copy()) 